backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.0
drf-spectacular==0.27.0
pillow==11.1.0
drf-yasg==1.21.10

# Optional: Required only for PostgreSQL database
psycopg2-binary==2.9.9

# to helps manage hierarchical data (nested categories)
django-ltree==0.5.3

# Image processing
django-imagekit==4.1.0

# Stripe
stripe==11.6.0
python-decouple==3.8

# Redis for caching
redis==5.2.1
django-redis==5.4.0

# CSV processing
python-slugify==8.0.4

# Excel processing
XlsxWriter==3.2.2
pandas==2.2.3

# Google Maps API
googlemaps==4.10.0

# Full-text search
djangorestframework-gis==1.1

# Celery for task queue and event processing
celery==5.3.6
django-celery-results==2.5.1
django-celery-beat==2.7.0
amqp==5.2.0

# Firebase Cloud Messaging
firebase-admin==6.7.0

# WebSockets. Real-time messaging system
channels==4.2.2
channels-redis==4.2.1  # Redis as our channel layer
uvicorn==0.34.0  # ASGI server for use with Gunicorn
gunicorn==23.0.0 # WSGI HTTP Server for UNIX with Uvicorn workers

# for entities to json
pydantic==2.11.2